Prevent trying to write to the characteristic, when connection is lost, just stop navigating

+++ Common/Navigation/Messages/DistMessage.cs
@@ -2,9 +2,9 @@
 
 public class DistMessage : BleMessage
 {
-    public float Distance { get; }
+    public uint Distance { get; }
 
-    public DistMessage(float distance): base(MessageType.DIST)
+    public DistMessage(uint distance): base(MessageType.DIST)
     {
         Distance = distance;
